Separated queue family indices struct
